requirements.txt
PyQt6>=6.5.0
pypdf>=3.15.0
pymupdf>=1.22.0
Pillow>=10.0.0
pyHanko>=0.20.0
